Add soundcloud_api.rb; can make external soundcloud API calls [#21]
The official soundcloud gem is great and has a lot of functionality but
it seemed a little heavyweight for our use case.  So, here is a small
client that retrieves favorited tracks for users.  Right now that's how
we're using soundcloud, pulling in tracks that a user favorites on
soundcloud into their tunefish feed.