Can get tweets with #tf from twitter
So for testing I thought it would be difficult to use VCR because with
twitter, you get an oauth token/secret for a user and make reqeusts to
the api on their behalf.  They are passed into the TwitterAPI class, but
I left them blank in the tests and VCR was able to successfully capture
and save a cassette.  So, maybe that particular reqeust just required
the twitter client/secret?