Can get ombedded tweets from the twitter api
This was originally just for a single tweet, but I think it makes more
sense to just get back the entire collection.  Also not sure as to
whether it needs to be tested as a public method.  As far as the
interface, the most important method is the one that returns the oembed
html strings, which is what will go into the database (that's what gets
plugged in dynamically) for twitter activity.