Move saving additional provider info down to models
This part is likely subject to change for twitter once we start figuring out oauth
tokens and secrets for twitter.  It was a temporary workaround to make
the requests based on twitter nicknames, as those requests are covered
under general discovery (but not feasible long term, we need to
authenticate on behalf of a user).