Add view helpers to activities/index.hbs for filtering by provider
Also adds more specific classes for youtube/soundcloud activity in order
to be able to toggle them separately.  jQuery toggle() flips the display
property on the given elements (from block to none) for example.  There
are other ways to do it but it seems to be a good solution, it's super
responsive once all of the activities have been loaded.